Perspective project maybe

+++ constants.h
@@ -1,5 +1,5 @@
 #pragma once
 
-constexpr int SCREEN_WIDTH = 800;
-constexpr int SCREEN_HEIGHT = 600;
+constexpr double SCREEN_WIDTH = 800;
+constexpr double SCREEN_HEIGHT = 600;
 constexpr double PI = 3.14159;
